requirements.txt
streamlit==1.35.0
openai-whisper
ffmpeg-python
torch==2.1.0+cpu
